requirements.txt
fastapi
uvicorn[standard]
transformers
torch
Pillow
requests
pydantic
pydantic-settings
python-dotenv